more portability fixes.
made sure the slice number is always passed as an integer, instead of
sometimes int and smetimes uint8_t

do not include mmx.h when on non-X86 builds (because of the nonstandard
long long in there)

+++ mpeg2dec/libmpeg2/slice.c
@@ -1505,7 +1505,7 @@ do {									\
     }									\
 } while (0)
 
-void slice_process (picture_t * picture, uint8_t code, uint8_t * buffer)
+void slice_process (picture_t * picture, int code, uint8_t * buffer)
 {
 #define bit_buf (picture->bitstream_buf)
 #define bits (picture->bitstream_bits)
